+++ CHANGELOG.md
@@ -5,6 +5,7 @@ chef_work_environment cookbook.
 
 ## v0.2.0 - 10.03.2021
 
+- Add example policyfile for use
 - Add configurability for Omnitruck, Chef Server/Supermarket and Rubygems
 - Add more Chef environment variables
 - Add global Chef configuration file
@@ -12,6 +13,7 @@ chef_work_environment cookbook.
 - Add tagging to `gitlab_runner`
 - Remove `gitlab_runner` from `default` recipe
 - Update `gitlab_runner` recipe to v13.9.0
+- Fix chef_workstation version to work on Debian/Ubuntu as well
 - Fix config file parameter on `gitlab_runner`
 - Fix executor type on `gitlab_runner`